requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=0.15.0=pyhd3eb1b0_0
attrs=21.4.0=pypi_0
blas=1.0=mkl
blinker=1.4=py36h06a4308_0
boto3=1.20.28=pypi_0
botocore=1.23.28=pypi_0
bravado=11.0.3=pypi_0
bravado-core=5.17.0=pypi_0
brotlipy=0.7.0=py36h27cfd23_1003
c-ares=1.17.1=h27cfd23_0
ca-certificates=2022.6.15=ha878542_0
cachetools=2.0.0=py36_0
certifi=2020.6.20=pyhd3eb1b0_3
cffi=1.15.0=pypi_0
charset-normalizer=2.0.10=pypi_0
click=8.0.3=pypi_0
cloudpickle=2.0.0=pyhd3eb1b0_0
coverage=5.5=py36h27cfd23_2
cryptography=3.4.8=py36hb60f036_0
cuda90=1.0=h6433d27_0
cudatoolkit=9.2=0
cycler=0.11.0=pyhd3eb1b0_0
cytoolz=0.11.0=py36h7b6447c_0
dask-core=1.1.4=py36_1
dataclasses=0.8=pyh4f3eec9_6
decorator=5.1.0=pyhd3eb1b0_0
dotmap=1.3.30=pypi_0
easydict=1.9=pypi_0
faiss-gpu=1.7.2=pypi_0
freetype=2.11.0=h70c0345_0
future=0.18.2=pypi_0
gitdb=4.0.9=pypi_0
gitpython=3.1.18=pypi_0
google-auth=1.35.0=pypi_0
google-auth-oauthlib=0.4.1=py_2
gputil=1.4.0=pypi_0
grpcio=1.35.0=py36hce63b2e_0
icu=67.1=he1b5a44_0
idna=3.3=pyhd3eb1b0_0
imageio=2.13.5=pypi_0
importlib-metadata=4.8.3=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
jmespath=0.10.0=pypi_0
joblib=1.1.0=pypi_0
jpeg=9d=h7f8727e_0
jsonpointer=2.2=pypi_0
jsonref=0.2=pypi_0
jsonschema=3.2.0=pypi_0
kiwisolver=1.3.1=py36h2531618_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libblas=3.9.0=12_linux64_mkl
libcblas=3.9.0=12_linux64_mkl
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=9.3.0=ha5ec8a7_17
libgfortran5=9.3.0=ha5ec8a7_17
libgomp=9.3.0=h5101ec6_17
liblapack=3.9.0=12_linux64_mkl
libpng=1.6.37=hbc83047_0
libprotobuf=3.17.2=h4ff587b_1
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.2.0=h85742a9_0
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.0=h27cfd23_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=py36h06a4308_0
matplotlib=3.2.2=1
matplotlib-base=3.2.2=py36h5fdd944_1
mkl=2021.4.0=h06a4308_640
monotonic=1.6=pypi_0
msgpack=1.0.3=pypi_0
multicoretsne=0.1=pypi_0
ncurses=6.3=h7f8727e_2
neptune-client=0.14.0=pypi_0
networkx=2.2=py36_1
ninja=1.10.2=h5e70eb0_2
numpy=1.19.5=py36hfc0c790_2
oauthlib=3.1.1=pyhd3eb1b0_0
olefile=0.46=py36_0
opencv-contrib-python=4.6.0.66=pypi_0
openssl=1.1.1q=h7f8727e_0
packaging=21.3=pypi_0
pandas=1.1.5=pypi_0
patsy=0.5.2=pyhd8ed1ab_0
pillow=8.3.1=pypi_0
pip=21.2.2=py36h06a4308_0
protobuf=3.17.2=py36h295c915_0
psutil=5.9.0=pypi_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pyjwt=2.3.0=pypi_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=3.0.4=pyhd3eb1b0_0
pyrsistent=0.18.0=pypi_0
pysocks=1.7.1=py36h06a4308_0
python=3.6.13=h12debd9_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.6=2_cp36m
pytorch=1.7.0=py3.6_cuda9.2.148_cudnn7.6.3_0
pytz=2021.3=pypi_0
pywavelets=1.1.1=py36h7b6447c_2
pyyaml=5.4.1=pypi_0
readline=8.1=h27cfd23_0
requests=2.27.0=pypi_0
requests-oauthlib=1.3.0=py_0
rfc3987=1.3.8=pypi_0
rsa=3.4.2=py_1
s3transfer=0.5.0=pypi_0
scikit-image=0.17.2=py36h284efc9_4
scikit-learn=0.24.2=pypi_0
scipy=1.5.4=pypi_0
seaborn=0.11.2=hd8ed1ab_0
seaborn-base=0.11.2=pyhd8ed1ab_0
setuptools=58.0.4=py36h06a4308_0
simplejson=3.17.6=pypi_0
six=1.16.0=pyhd3eb1b0_0
smmap=5.0.0=pypi_0
sqlite=3.37.0=hc218d9a_0
statsmodels=0.9.0=py36h3010b51_1000
strict-rfc3339=0.7=pypi_0
swagger-spec-validator=2.7.4=pypi_0
tensorboard=2.6.0=py_1
tensorboard-data-server=0.6.0=py36hca6d32c_0
tensorboard-plugin-wit=1.6.0=py_0
tensorboardx=2.5.1=pypi_0
threadpoolctl=3.0.0=pypi_0
tifffile=2020.10.1=py36hdd07704_2
timm=0.5.4=pypi_0
tk=8.6.11=h1ccaba5_0
toolz=0.11.2=pyhd3eb1b0_0
torch-tb-profiler=0.3.1=pypi_0
torchaudio=0.7.0=py36
torchvision=0.8.0=py36_cu92
tornado=6.1=py36h27cfd23_0
tqdm=4.62.3=pypi_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.7=pypi_0
webcolors=1.11.1=pypi_0
websocket-client=1.2.3=pypi_0
werkzeug=2.0.2=pyhd3eb1b0_0
wheel=0.37.0=pyhd3eb1b0_1
xz=5.2.5=h7b6447c_0
zipp=3.6.0=pyhd3eb1b0_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
